Adiciona seeds e garante filtragem por todos os status de solicitação
